+++ src/Validation/dtos/index.ts
@@ -26,3 +26,4 @@ export { IValidationIsUUIDDto } from "./validation.isuuid.dto";
 export { IValidationIsDateDto } from "./validation.isdate.dto";
 export { IValidationIsValidMongoIdDto } from "./validation.isvalidmongoid.dto";
 export { IValidationUniqueDto } from "./validation.unique.dto";
+export { IValidationRequiredWithKeysDto } from "./validation.requiredwithkeys.dto";